requirements.txt
sklearn
numpy
h5py
keras
tensorflow-gpu
librosa
matplotlib
seaborn
pandas
tqdm